requirements.txt
asgiref==3.11.1
Django==6.0.3
django-cors-headers==4.9.0
django-jazzmin==3.0.4
djangorestframework==3.17.1
pillow==12.1.1
sqlparse==0.5.5
